+++ src/oatpp/parser/json/mapping/ObjectMapper.cpp
@@ -26,15 +26,15 @@
 
 namespace oatpp { namespace parser { namespace json { namespace mapping {
 
-ObjectMapper::ObjectMapper(const std::shared_ptr<Serializer2::Config>& pSerializerConfig,
+ObjectMapper::ObjectMapper(const std::shared_ptr<Serializer::Config>& pSerializerConfig,
                            const std::shared_ptr<Deserializer::Config>& pDeserializerConfig)
   : data::mapping::ObjectMapper(getMapperInfo())
-  , m_serializer(std::make_shared<Serializer2>(pSerializerConfig))
+  , m_serializer(std::make_shared<Serializer>(pSerializerConfig))
   , serializerConfig(pSerializerConfig)
   , deserializerConfig(pDeserializerConfig)
 {}
 
-std::shared_ptr<ObjectMapper> ObjectMapper::createShared(const std::shared_ptr<Serializer2::Config>& serializerConfig,
+std::shared_ptr<ObjectMapper> ObjectMapper::createShared(const std::shared_ptr<Serializer::Config>& serializerConfig,
                                                          const std::shared_ptr<Deserializer::Config>& deserializerConfig){
   return std::make_shared<ObjectMapper>(serializerConfig, deserializerConfig);
 }
