--- src/oatpp/parser/json/mapping/Serializer2.hpp
@@ -1,166 +1,0 @@
-/***************************************************************************
- *
- * Project         _____    __   ____   _      _
- *                (  _  )  /__\ (_  _)_| |_  _| |_
- *                 )(_)(  /(__)\  )( (_   _)(_   _)
- *                (_____)(__)(__)(__)  |_|    |_|
- *
- *
- * Copyright 2018-present, Leonid Stryzhevskyi <lganzzzo@gmail.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- ***************************************************************************/
-
-#ifndef oatpp_parser_json_mapping_Serializer2_hpp
-#define oatpp_parser_json_mapping_Serializer2_hpp
-
-#include "oatpp/parser/json/Beautifier.hpp"
-
-#include "oatpp/core/data/mapping/type/ListMap.hpp"
-#include "oatpp/core/data/mapping/type/List.hpp"
-#include "oatpp/core/data/mapping/type/Object.hpp"
-#include "oatpp/core/data/mapping/type/Primitive.hpp"
-#include "oatpp/core/data/mapping/type/Type.hpp"
-#include "oatpp/core/data/stream/ChunkedBuffer.hpp"
-
-#include "oatpp/core/parser/Caret.hpp"
-
-#include "oatpp/core/collection/LinkedList.hpp"
-#include "oatpp/core/Types.hpp"
-
-#include <vector>
-
-namespace oatpp { namespace parser { namespace json { namespace mapping {
-
-/**
- * Json Serializer.
- * Serializes oatpp DTO object to json. See [Data Transfer Object(DTO) component](https://oatpp.io/docs/components/dto/).
- */
-class Serializer2 {
-public:
-  typedef oatpp::data::mapping::type::Type Type;
-  typedef oatpp::data::mapping::type::Type::Property Property;
-  typedef oatpp::data::mapping::type::Type::Properties Properties;
-
-  typedef oatpp::data::mapping::type::Object Object;
-  typedef oatpp::String String;
-
-  template<class T>
-  using PolymorphicWrapper = data::mapping::type::PolymorphicWrapper<T>;
-
-  typedef oatpp::data::mapping::type::AbstractObjectWrapper AbstractObjectWrapper;
-  typedef oatpp::data::mapping::type::List<AbstractObjectWrapper> AbstractList;
-  typedef oatpp::data::mapping::type::ListMap<String, AbstractObjectWrapper> AbstractFieldsMap;
-
-public:
-  /**
-   * Serializer config.
-   */
-  class Config : public oatpp::base::Countable {
-  public:
-    /**
-     * Constructor.
-     */
-    Config()
-    {}
-  public:
-
-    /**
-     * Create shared config.
-     * @return - `std::shared_ptr` to Config.
-     */
-    static std::shared_ptr<Config> createShared(){
-      return std::make_shared<Config>();
-    }
-
-    /**
-     * Include fields with value == nullptr into serialized json.
-     */
-    bool includeNullFields = true;
-
-    /**
-     * If `true` - insert string `"<unknown-type>"` in json field value in case unknown field found.
-     * Fail if `false`.
-     * Known types for this serializer are:<br>
-     * (String, Int8, Int16, Int32, Int64, Float32, Float64, Boolean, DTOs, List, Fields).
-     */
-    bool throwOnUnknownTypes = true;
-
-    /**
-     * Use JSON Beautifier.
-     */
-    bool useBeautifier = false;
-
-    /**
-     * Beautifier Indent.
-     */
-    oatpp::String beautifierIndent = "  ";
-
-    /**
-     * Beautifier new line.
-     */
-    oatpp::String beautifierNewLine = "\n";
-
-  };
-public:
-  typedef void (*SerializerMethod)(Serializer2*,
-                                   data::stream::ConsistentOutputStream*,
-                                   const data::mapping::type::AbstractObjectWrapper&);
-private:
-
-  template<class T>
-  static void serializePrimitive(Serializer2* serializer,
-                                 data::stream::ConsistentOutputStream* stream,
-                                 const data::mapping::type::AbstractObjectWrapper& polymorph){
-    if(polymorph){
-      stream->writeAsString(static_cast<typename T::ObjectType*>(polymorph.get())->getValue());
-    } else {
-      stream->writeSimple("null", 4);
-    }
-  }
-
-  static void serializeString(oatpp::data::stream::ConsistentOutputStream* stream, p_char8 data, v_buff_size size);
-  static void serializeString(Serializer2* serializer,
-                              data::stream::ConsistentOutputStream* stream,
-                              const data::mapping::type::AbstractObjectWrapper& polymorph);
-
-  static void serializeList(Serializer2* serializer,
-                            data::stream::ConsistentOutputStream* stream,
-                            const data::mapping::type::AbstractObjectWrapper& polymorph);
-
-  static void serializeFieldsMap(Serializer2* serializer,
-                                 data::stream::ConsistentOutputStream* stream,
-                                 const data::mapping::type::AbstractObjectWrapper& polymorph);
-
-  static void serializeObject(Serializer2* serializer,
-                              data::stream::ConsistentOutputStream* stream,
-                              const data::mapping::type::AbstractObjectWrapper& polymorph);
-private:
-  std::vector<SerializerMethod> m_methods;
-  std::shared_ptr<Config> m_config;
-public:
-
-  Serializer2(const std::shared_ptr<Config>& config = std::make_shared<Config>());
-
-  void setSerializerMethod(const data::mapping::type::ClassId& classId, SerializerMethod method);
-
-  void serialize(data::stream::ConsistentOutputStream* stream, const data::mapping::type::AbstractObjectWrapper& polymorph);
-
-  const std::shared_ptr<Config>& getConfig();
-
-};
-
-}}}}
-
-#endif /* oatpp_parser_json_mapping_Serializer2_hpp */
